renamed namespaces and files
